+++ CHANGELOG.rst
@@ -34,6 +34,8 @@ see TODO.md for list of TODO items
 
   * changed parameter names on Regex cleaner to match re.sub parameter names
 
+  * added count parameter to RemoveCleaner
+
   * removed TableConvertor. This functionality has been replaced by the Table class.
 
 * v0.2.12: